+++ lib/new_window_method_channel.dart
@@ -11,6 +11,9 @@ class MethodChannelNewWindow extends NewWindowPlatform {
   @visibleForTesting
   final methodChannel = const MethodChannel('new_window');
 
+  @visibleForTesting
+  final eventChannel = const EventChannel('new_window_event');
+
   @override
   Future<String?> getPlatformVersion() async {
     final version =
@@ -20,7 +23,17 @@ class MethodChannelNewWindow extends NewWindowPlatform {
 
   @override
   Future<int> createWindow({Rect? rect, bool closable = true}) async {
-    final windowId = await methodChannel.invokeMethod<int>('createWindow');
+    var arguments = <String, dynamic>{
+      'closable': closable,
+    };
+    if (rect != null) {
+      arguments['px'] = rect.left;
+      arguments['py'] = rect.top;
+      arguments['width'] = rect.size.width;
+      arguments['height'] = rect.size.height;
+    }
+    final windowId =
+        await methodChannel.invokeMethod<int>('createWindow', arguments);
     if (windowId == null) {
       return Future.error(
           FlutterError('Create window failed with null window id'));
@@ -44,5 +57,23 @@ class MethodChannelNewWindow extends NewWindowPlatform {
     };
 
     return methodChannel.invokeMethod('showWindow', arguments);
+  }
+
+  @override
+  Stream receiver() => eventChannel.receiveBroadcastStream();
+
+  @override
+  Future<bool?> sendMessage(
+      {required int fromWindowId,
+      required int toWindwoId,
+      required String message}) {
+    return methodChannel.invokeMethod<bool>(
+      'sendMessage',
+      {
+        'fromWindowId': fromWindowId,
+        'toWindowId': toWindwoId,
+        'message': message,
+      },
+    );
   }
 }