+++ macos/Classes/NewWindow.swift
@@ -11,18 +11,19 @@ import Foundation
 class NewWindow: NSObject, NSWindowDelegate {
     private static var globalWindowsId: Int = 0
     
-    private let window: NSWindow = .init(contentRect: NSRect(x: 0, y: 0, width: 1024, height: 768), styleMask: [.miniaturizable, .closable, .resizable, .titled, .fullSizeContentView], backing: .buffered, defer: false)
-    
+    private let window: NSWindow = .init(contentRect: NSRect(x: 0, y: 0, width: 1024, height: 768), styleMask: [.miniaturizable, .closable, .resizable, .titled], backing: .buffered, defer: false)
+      
     private lazy var windowController = NSWindowController(window: window)
-    
     private lazy var id: Int = {
         NewWindow.globalWindowsId += 1
         return NewWindow.globalWindowsId
     }()
     
-    private var closable: Bool = true
+    private let windowClosable: Bool
     
-    private var rect: NSRect = .init(x: 0, y: 0, width: 1024, height: 768)
+    private let initialPoint: NSPoint?
+    
+    private let initialSize: NSSize?
     
     private var eventChannel: NewWindowEventChannel?
     
@@ -33,19 +34,16 @@ class NewWindow: NSObject, NSWindowDelegate {
     open weak var delegate: NewWindowDelegate?
     
     // return the window id
-    init(rect: NSRect?, closable: Bool?) {
+    init(point: NSPoint?, size: NSSize?, closable: Bool?) {
+        windowClosable = closable ?? true
+        initialPoint = point
+        initialSize = size
         super.init()
-        self.closable = closable ?? true
-        self.rect = rect ?? NSRect(x: 0, y: 0, width: 1024, height: 768)
         window.delegate = self
-        window.isReleasedWhenClosed = true
     }
     
     deinit {
         delegate = nil
-        if let flutterViewController = windowController.contentViewController as? FlutterViewController {
-            flutterViewController.engine.shutDownEngine()
-        }
     }
     
     // open the window
@@ -61,35 +59,60 @@ class NewWindow: NSObject, NSWindowDelegate {
         
         NewWindowPlugin.register(with: plugin)
         
+        // show in the main window's screen
+        let mainWindowScreen = NSApp.mainWindow?.screen ?? NSScreen.main
+        let screenPoint = mainWindowScreen?.frame.origin ?? NSPoint.zero
+       
+        var frameOrigin = initialPoint ?? NSPoint.zero
+        frameOrigin = NSPoint(x: frameOrigin.x + screenPoint.x, y: frameOrigin.y + screenPoint.y)
+        let frameSize = initialSize ?? NSApp.mainWindow?.frame.size ?? NSSize(width: 1024, height: 760)
+        
+        let showRect = NSRect(origin: frameOrigin, size: frameSize)
         window.contentViewController = viewController
-        window.contentRect(forFrameRect: rect)
-        window.setFrame(rect, display: true, animate: true)
+        window.contentRect(forFrameRect: showRect)
+        window.setFrame(showRect, display: true, animate: true)
+        
+        if initialPoint == nil {
+            window.center()
+        }
+//        window.makeKeyAndOrderFront(self)
         
         windowController.contentViewController = window.contentViewController
         windowController.shouldCascadeWindows = true
         windowController.showWindow(self)
-//        window.makeKeyAndOrderFront(nil)
-//        NSApp.activate(ignoringOtherApps: true)
     }
     
     // close the window
     func close() {
+//        window.orderOut(self)
+//        window.close()
         windowController.close()
+        window.close()
+    }
+    
+    func sendMessage(fromWindowId: Int, message: String) -> Bool {
+        return eventChannel?.emit(withState: NewWindowState.onMessage, andData: ["fromWindowId": fromWindowId, "message": message]) ?? false
     }
     
     // NSWindowDelegate
     func windowWillClose(_ notification: Notification) {
+        _ = eventChannel?.emit(withState: NewWindowState.onClose, andData: nil)
+        if let flutterViewController = window.contentViewController as? FlutterViewController {
+            // 延时释放Engine资源
+            DispatchQueue.main.asyncAfter(deadline: DispatchTime.now() + 1) {
+                flutterViewController.engine.shutDownEngine()
+            }
+        }
         delegate?.windowWillClose(id: id)
-        _ = eventChannel?.emit(withWindowId: id, andState: NewWindowState.onClose)
     }
-    
+   
     func windowShouldClose(_ sender: NSWindow) -> Bool {
-        let ret = eventChannel?.emit(withWindowId: id, andState: NewWindowState.shouldClose) ?? false
+        let ret = eventChannel?.emit(withState: NewWindowState.shouldClose, andData: nil) ?? false
         if !ret {
             return true
         }
         
-        return closable
+        return windowClosable
     }
 }
 
@@ -100,6 +123,7 @@ public protocol NewWindowDelegate: NSObjectProtocol {
 enum NewWindowState {
     case shouldClose
     case onClose
+    case onMessage
     
     func toString() -> String {
         switch self {
@@ -107,6 +131,8 @@ enum NewWindowState {
             return "shouldClose"
         case .onClose:
             return "willClose"
+        case .onMessage:
+            return "onMessage"
         }
     }
 }