+++ macos/Classes/NewWindowController.swift
@@ -11,8 +11,8 @@ import Foundation
 class NewWindowController: NSObject, NewWindowDelegate {
     private static var windows = [Int: NewWindow]()
    
-    func createWindow(withRect rect: NSRect?, closable: Bool?) -> NewWindow {
-        let window = NewWindow(rect: rect, closable: closable)
+    func createWindow(withPoint point: NSPoint?, andSize size: NSSize?, closable: Bool?) -> NewWindow {
+        let window = NewWindow(point: point, size: size, closable: closable)
         window.delegate = self
         NewWindowController.windows[window.windowId] = window
         
@@ -27,11 +27,15 @@ class NewWindowController: NSObject, NewWindowDelegate {
     func closeWindow(id: Int) {
         let window = NewWindowController.windows[id]
         window?.close()
-        NewWindowController.windows.removeValue(forKey: id)
     }
     
     func isWindowExist(id: Int) -> Bool {
         return NewWindowController.windows[id] != nil
+    }
+    
+    func sendMessage(fromWindowId: Int, toWindowId: Int, message: String) {
+        let window = NewWindowController.windows[toWindowId]
+        window?.sendMessage(fromWindowId: fromWindowId, message: message)
     }
     
     // NewWindowDelegate