+++ macos/Classes/NewWindowPlugin.swift
@@ -16,66 +16,70 @@ public class NewWindowPlugin: NSObject, FlutterPlugin {
             result("macOS " + ProcessInfo.processInfo.operatingSystemVersionString)
         case "createWindow":
             guard let arguments = call.arguments as? [String: Any]? else {
-                let error = NewWindowError.argumentError(message: "CreateWindow Arguments: arguments must be Dictionary")
+                let error = NewWindowError.argumentError(message: "createWindow Arguments: arguments must be Dictionary")
                 result(FlutterError(code: error.code, message: error.message, details: nil))
                 return
             }
             guard let x = arguments?["px"] as? Double? else {
-                let error = NewWindowError.argumentError(message: "CreateWindow Arguments: The field 'px' must be Double")
+                let error = NewWindowError.argumentError(message: "createWindow Arguments: The field 'px' must be Double")
                 result(FlutterError(code: error.code, message: error.message, details: nil))
                 return
             }
 
             guard let y = arguments?["py"] as? Double? else {
-                let error = NewWindowError.argumentError(message: "CreateWindow Arguments: The field 'py' must be Double")
+                let error = NewWindowError.argumentError(message: "createWindow Arguments: The field 'py' must be Double")
                 result(FlutterError(code: error.code, message: error.message, details: nil))
                 return
             }
 
             guard let width = arguments?["width"] as? Double? else {
-                let error = NewWindowError.argumentError(message: "CreateWindow Arguments: The field 'width' must be Double")
+                let error = NewWindowError.argumentError(message: "createWindow Arguments: The field 'width' must be Double")
                 result(FlutterError(code: error.code, message: error.message, details: nil))
                 return
             }
 
             guard let height = arguments?["height"] as? Double? else {
-                let error = NewWindowError.argumentError(message: "CreateWindow Arguments: The field 'height' must be Double")
+                let error = NewWindowError.argumentError(message: "createWindow Arguments: The field 'height' must be Double")
                 result(FlutterError(code: error.code, message: error.message, details: nil))
                 return
             }
 
             guard let closable = arguments?["closable"] as? Bool? else {
-                let error = NewWindowError.argumentError(message: "CreateWindow Arguments: The field 'closable' must be Bool")
+                let error = NewWindowError.argumentError(message: "createWindow Arguments: The field 'closable' must be Bool")
                 result(FlutterError(code: error.code, message: error.message, details: nil))
                 return
             }
-
-            let position = NSPoint(x: x ?? 0, y: y ?? 0)
-            let size = NSSize(width: width ?? 1024, height: height ?? 768)
-
-            let window = windowController.createWindow(withRect: NSRect(origin: position, size: size), closable: closable)
+            var point: NSPoint?
+            if x != nil, y != nil {
+                point = NSPoint(x: x!, y: y!)
+            }
+            var size: NSSize?
+            if width != nil, height != nil {
+                size = NSSize(width: width!, height: height!)
+            }
+            let window = windowController.createWindow(withPoint: point, andSize: size, closable: closable)
             result(window.windowId)
         case "showWindow":
             guard let arguments = call.arguments as? [String: Any]? else {
-                let error = NewWindowError.argumentError(message: "ShowWindow Arguments: arguments must be Dictionary")
+                let error = NewWindowError.argumentError(message: "showWindow Arguments: arguments must be Dictionary")
                 result(FlutterError(code: error.code, message: error.message, details: nil))
                 return
             }
 
             guard let windowId = arguments?["windowId"] as? Int else {
-                let error = NewWindowError.argumentError(message: "ShowWindow Arguments: Lack the 'windowId' field or not Int type")
+                let error = NewWindowError.argumentError(message: "showWindow Arguments: Lack the 'windowId' field or not Int type")
                 result(FlutterError(code: error.code, message: error.message, details: nil))
                 return
             }
 
             guard let route = arguments?["route"] as? String? else {
-                let error = NewWindowError.argumentError(message: "ShowWindow Arguments: The field 'route' must be String")
+                let error = NewWindowError.argumentError(message: "showWindow Arguments: The field 'route' must be String")
                 result(FlutterError(code: error.code, message: error.message, details: nil))
                 return
             }
 
             guard let windowArgs = arguments?["args"] as? String? else {
-                let error = NewWindowError.argumentError(message: "ShowWindow Arguments: The field 'args' must be String")
+                let error = NewWindowError.argumentError(message: "showWindow Arguments: The field 'args' must be String")
                 result(FlutterError(code: error.code, message: error.message, details: nil))
                 return
             }
@@ -84,13 +88,35 @@ public class NewWindowPlugin: NSObject, FlutterPlugin {
             result(nil)
         case "closeWindow":
             guard let windowId = call.arguments as? Int else {
-                let error = NewWindowError.argumentError(message: "CloseWindow need a window id")
+                let error = NewWindowError.argumentError(message: "closeWindow need a window id")
                 result(FlutterError(code: error.code, message: error.message, details: nil))
                 return
             }
-
-            windowController.closeWindow(id: windowId)
             result(nil)
+            windowController.closeWindow(id: windowId)
+        case "sendMessage":
+            guard let arguments = call.arguments as? [String: Any] else {
+                let error = NewWindowError.argumentError(message: "sendMessage Arguments: arguments must be Dictionary!")
+                result(FlutterError(code: error.code, message: error.message, details: nil))
+                return
+            }
+            guard let fromWindowId = arguments["fromWindowId"] as? Int else {
+                let error = NewWindowError.argumentError(message: "sendMessage Arguments: The field 'fromWindowId' must be Int!")
+                result(FlutterError(code: error.code, message: error.message, details: nil))
+                return
+            }
+            guard let toWindowId = arguments["toWindowId"] as? Int else {
+                let error = NewWindowError.argumentError(message: "sendMessage Arguments: The field 'toWindowId' must be Int!")
+                result(FlutterError(code: error.code, message: error.message, details: nil))
+                return
+            }
+            guard let message = arguments["message"] as? String else {
+                let error = NewWindowError.argumentError(message: "sendMessage Arguments: The field 'message' must be String!")
+                result(FlutterError(code: error.code, message: error.message, details: nil))
+                return
+            }
+            windowController.sendMessage(fromWindowId: fromWindowId, toWindowId: toWindowId, message: message)
+
         default:
             result(FlutterMethodNotImplemented)
         }