+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Basic operation:
+  1. Open a new window.
+  2. Close the window
+  3. Listen the new window on closed.
+* Support in MacOS.